Move repo search to utils
