+++ integrations/github/github/helpers/utils.py
@@ -1,7 +1,20 @@
 from enum import StrEnum
-from typing import Any, Dict, List, NamedTuple, Optional, Set, Tuple
+from typing import (
+    Any,
+    AsyncGenerator,
+    Dict,
+    Iterable,
+    List,
+    NamedTuple,
+    Optional,
+    Set,
+    Tuple,
+)
 from loguru import logger
 from typing import TYPE_CHECKING
+
+from port_ocean.utils.async_iterators import stream_async_iterators_tasks
+from port_ocean.utils.cache import cache_iterator_result
 
 if TYPE_CHECKING:
     from github.clients.http.base_client import AbstractGithubClient
@@ -99,6 +112,73 @@ def enrich_with_commit(
     """Helper function to enrich response with commit information."""
     response["commit"] = commit_object
     return response
+
+
+def create_search_params(repos: Iterable[str], max_operators: int = 5) -> list[str]:
+    """Create search query strings that fits into Github search string limitations.
+
+    Limitations:
+        - A search query can be up to 256 characters.
+        - A query can contain a maximum of 5 `OR` operators.
+
+    """
+    search_strings = []
+    if not repos:
+        return []
+
+    max_repos_in_query = max_operators + 1
+    max_search_string_len = 256
+
+    chunk: list[str] = []
+    current_query = ""
+    for repo in repos:
+        repo_query_part = f"{repo} in:name"
+
+        if len(repo_query_part) > max_search_string_len:
+            logger.warning(
+                f"Repository name '{repo}' is too long to fit in a search query."
+            )
+            continue
+
+        if not chunk:
+            chunk.append(repo)
+            current_query = repo_query_part
+            continue
+
+        if (
+            len(chunk) + 1 > max_repos_in_query
+            or len(f"{current_query} OR {repo_query_part}") > max_search_string_len
+        ):
+            search_strings.append(current_query)
+            chunk = [repo]
+            current_query = repo_query_part
+        else:
+            chunk.append(repo)
+            current_query = f"{current_query} OR {repo_query_part}"
+
+    if chunk:
+        search_strings.append(current_query)
+
+    return search_strings
+
+
+@cache_iterator_result()
+async def search_for_repositories(
+    client: "AbstractGithubClient", repos: Iterable[str]
+) -> AsyncGenerator[list[dict[str, Any]], None]:
+    """Search Github for a list of repositories and cache the result"""
+
+    tasks = []
+    for search_string in create_search_params(repos):
+        logger.debug(f"creating a search task for search string: {search_string}")
+        query = f"org:{client.organization} {search_string} fork:true"
+        url = f"{client.base_url}/search/repositories"
+        params = {"q": query}
+        tasks.append(client.send_paginated_request(url, params=params))
+
+    async for search_result in stream_async_iterators_tasks(*tasks):
+        valid_repos = [repo for repo in search_result["items"] if repo["name"] in repos]
+        yield valid_repos
 
 
 class IgnoredError(NamedTuple):
